requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
faiss-cpu>=1.9.0
numpy>=1.26.0
huggingface_hub>=0.24.0
requests>=2.32.0
pdfplumber>=0.11.0
python-multipart>=0.0.12
streamlit>=1.35.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
